backend/dashboard/requirements.txt
fastapi>=0.116.1
uvicorn>=0.35.0
python-dotenv>=1.0.0